pyproject.toml
[tool.poetry]
name = "project-etl"
version = "0.1.0"
description = ""
authors = ["othon <othonbrener123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
sqlalchemy = "^2.0.31"
fastapi = "^0.111.1"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
pandas = "^2.2.2"
isort = "^5.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.2"
ipython = "^8.26.0"

[tool.ruff]
line-length = 79
extend-exclude = ["migration"]

[tool.ruff.lint]
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.isort]
multi_line_output = 3
line_length = 79
include_trailing_comma = true

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev project/api.py'
test = 'pytest -s -x --cov=project -vv'
post_test = 'coverage html'
init_db = 'docker compose up -d && cd project/orm_fonte && alembic upgrade head'
down_db = 'docker compose down --remove-orphans --volumes'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
